move Parser trait to utils
this allow the parser capabilities to be re-used for json module
